+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - CIP Connection Slot can now be a number, a string, or a buffer
 - CIP Identity added Device Type Names
 - CIP Connection added decoding for instance attributes
+- CIP Connection/ConnectionManager now support Large Forward Open
 
 ## 2.0.0-beta.3 (2019-12-7)
 - `Logix5000.readTag()` now reads the entire array if the tag is a 1-dimensional array